+++ src/utils.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 ISciences, LLC.
+// Copyright (c) 2019-2024 ISciences, LLC.
 // All rights reserved.
 //
 // This software is licensed under the Apache License, Version 2.0 (the "License").
@@ -12,10 +12,12 @@
 // limitations under the License.
 
 #include "utils.h"
+#include "operation.h"
+#include "raster_source.h"
 
 #include <regex>
-#include <sstream>
 #include <stdexcept>
+#include <unordered_map>
 
 namespace exactextract {
 
@@ -58,19 +60,14 @@ parse_raster_descriptor(const std::string& descriptor)
     }
 
     if (pos2 == std::string::npos) {
-        // No band was specified; set it to 1.
+        // No band was specified; set it to 0.
         fname = descriptor.substr(pos1 + 1);
-        band = 1;
+        band = 0;
     } else {
         fname = descriptor.substr(pos1 + 1, pos2 - pos1 - 1);
 
         auto rest = descriptor.substr(pos2 + 1);
         band = std::stoi(rest);
-    }
-
-    if (pos1 == std::string::npos) {
-        // No name was provided, so just use the filename
-        name = fname;
     }
 
     if (fname.empty())
@@ -80,11 +77,13 @@ parse_raster_descriptor(const std::string& descriptor)
 }
 
 StatDescriptor
-parse_stat_descriptor(const std::string& descriptor)
+parse_stat_descriptor(const std::string& p_descriptor)
 {
-    if (descriptor.empty()) {
+    if (p_descriptor.empty()) {
         throw std::runtime_error("Invalid stat descriptor.");
     }
+
+    std::string descriptor = p_descriptor;
 
     StatDescriptor ret;
 
@@ -93,10 +92,11 @@ parse_stat_descriptor(const std::string& descriptor)
     std::smatch result_name_match;
     if (std::regex_search(descriptor, result_name_match, re_result_name)) {
         ret.name = result_name_match[1].str();
+        descriptor.erase(0, result_name_match.length());
     }
 
-    // Parse name of value raster
-    const std::regex re_func_name("=?(\\w+)\\(");
+    // Parse name of stat
+    const std::regex re_func_name("=?(\\w+)\\(?");
     std::smatch func_name_match;
     if (std::regex_search(descriptor, func_name_match, re_func_name)) {
         ret.stat = func_name_match[1].str();
@@ -104,7 +104,7 @@ parse_stat_descriptor(const std::string& descriptor)
         throw std::runtime_error("Invalid stat descriptor.");
     }
 
-    // Parse name of weight raster
+    // Parse stat arguments
     const std::regex re_args(R"(\(([,\w]+)+\)$)");
     std::smatch arg_names_match;
     if (std::regex_search(descriptor, arg_names_match, re_args)) {
@@ -117,23 +117,115 @@ parse_stat_descriptor(const std::string& descriptor)
             ret.values = args.substr(0, pos);
             ret.weights = args.substr(pos + 1);
         }
-    } else {
-        throw std::runtime_error("Invalid stat descriptor.");
-    }
-
-    // Construct a name if one was not specified
-    if (ret.name.empty()) {
-        std::ostringstream ss;
-        ss << ret.values << '_' << ret.stat;
-
-        if (!ret.weights.empty()) {
-            ss << '_' << ret.weights;
-        }
-
-        ret.name = ss.str();
     }
 
     return ret;
 }
 
+static std::string
+make_name(const RasterSource* v, const RasterSource* w, const std::string& stat, bool full_names)
+{
+    if (!full_names) {
+        return stat;
+    }
+
+    if (starts_with(stat, "weighted")) {
+        if (w == nullptr) {
+            throw std::runtime_error("No weights specified for stat: " + stat);
+        }
+
+        return v->name() + "_" + w->name() + "_" + stat;
+    }
+
+    return v->name() + "_" + stat;
+}
+
+static void
+prepare_operations_implicit(
+  std::vector<std::unique_ptr<Operation>>& ops,
+  const StatDescriptor& sd,
+  const RasterSourceVect& values,
+  const RasterSourceVect& weights)
+{
+    const bool full_names = values.size() > 1 || weights.size() > 1;
+
+    if (values.size() > 1 && weights.size() > 1 && values.size() != weights.size()) {
+        throw std::runtime_error("Value and weight rasters must have a single band or the same number of bands.");
+    }
+
+    for (std::size_t i = 0; i < values.size(); i++) {
+        RasterSource* v = &*values[i % values.size()];
+        RasterSource* w = weights.empty() ? nullptr : &*weights[i % weights.size()];
+
+        ops.push_back(std::make_unique<Operation>(
+          sd.stat,
+          make_name(v, w, sd.stat, full_names),
+          v,
+          w));
+    }
+}
+
+void
+prepare_operations_explicit(
+  std::vector<std::unique_ptr<Operation>>& ops,
+  const StatDescriptor& stat,
+  const RasterSourceVect& raster_sources,
+  const RasterSourceVect& weight_sources)
+{
+    std::unordered_map<std::string, RasterSource*> source_map;
+    std::unordered_map<std::string, RasterSource*> weights_map;
+
+    for (const auto& rast : raster_sources) {
+        source_map[rast->name()] = rast.get();
+        weights_map[rast->name()] = rast.get();
+    }
+
+    for (const auto& rast : weight_sources) {
+        weights_map[rast->name()] = rast.get();
+    }
+
+    auto values_it = source_map.find(stat.values);
+    if (values_it == source_map.end()) {
+        throw std::runtime_error("Unknown raster " + stat.values + " in stat " + stat.stat);
+    }
+
+    RasterSource* values = values_it->second;
+    RasterSource* weights = nullptr;
+
+    if (!stat.weights.empty()) {
+        auto weights_it = weights_map.find(stat.weights);
+        if (weights_it == weights_map.end()) {
+            weights_it = source_map.find(stat.weights);
+
+            if (weights_it == weights_map.end()) {
+                throw std::runtime_error("Unknown raster " + stat.weights + " in stat " + stat.stat);
+            }
+        }
+
+        weights = weights_it->second;
+    }
+
+    ops.emplace_back(std::make_unique<Operation>(stat.stat, stat.name.empty() ? values->name() + "_" + stat.stat : stat.name, values, weights));
+}
+
+std::vector<std::unique_ptr<Operation>>
+prepare_operations(
+  const std::vector<std::string>& descriptors,
+  const std::vector<std::unique_ptr<RasterSource>>& rasters,
+  const std::vector<std::unique_ptr<RasterSource>>& weights)
+{
+    std::vector<std::unique_ptr<Operation>> ops;
+
+    std::vector<StatDescriptor> parsed_descriptors;
+    for (const auto& descriptor : descriptors) {
+        auto parsed = parse_stat_descriptor(descriptor);
+        if (parsed.values.empty() && parsed.weights.empty()) {
+            prepare_operations_implicit(ops, parsed, rasters, weights);
+        } else {
+            prepare_operations_explicit(ops, parsed, rasters, weights);
+        }
+    }
+
+    return ops;
+}
 }